Fixed a lot of compilation warnings;

+++ Source/Core/Images/Utils.cpp
@@ -276,7 +276,8 @@ protected:
 int* boxesForGauss(float sigma, int n)  // standard deviation, number of boxes
 {
     float wIdeal = sqrt((12*sigma*sigma/n)+1);  // Ideal averaging filter width 
-    int wl = floor(wIdeal);  if(wl%2==0) wl--;
+    int wl = static_cast<int>(floor(wIdeal));  
+    if(wl%2==0) wl--;
     int wu = wl+2;
 
     float mIdeal = (12*sigma*sigma - n*wl*wl - 4*n*wl - 3*n)/(-4*wl - 4);
@@ -289,20 +290,21 @@ int* boxesForGauss(float sigma, int n)  // standard deviation, number of boxes
 }
 
 void boxBlurH_4 (DummyBitmap& scl, DummyBitmap& tcl, int w, int h, int r) {
-    float iarr = 1.0 / (r+r+1);
+    float iarr = static_cast<float>(1.0 / (r + r + 1));
     for(int i=0; i<h; i++) {
         int ti = i*w, li = ti, ri = ti+r;
         int fv = scl[ti], lv = scl[ti+w-1], val = (r+1)*fv;
         for(int j=0; j<r; j++) val += scl[ti+j];
         for(int j=0  ; j<=r ; j++) { 
             val += scl[ri++] - fv       ;   
-            tcl[ti++] = round(val*iarr); }
+            tcl[ti++] = static_cast<uint8_t>(round(val*iarr)); 
+        }
         for(int j=r+1; j<w-r; j++) { val += scl[ri++] - scl[li++];   tcl[ti++] = round(val*iarr); }
         for(int j=w-r; j<w  ; j++) { val += lv        - scl[li++];   tcl[ti++] = round(val*iarr); }
     }
 }
 void boxBlurT_4 (DummyBitmap&  scl, DummyBitmap&  tcl, int w, int h, int r) {
-    float iarr = 1.0 / (r+r+1);
+    float iarr = static_cast<float>(1.0 / (r+r+1));
     for(int i=0; i<w; i++) {
         int ti = i, li = ti, ri = ti+r*w;
         int fv = scl[ti], lv = scl[ti+w*(h-1)], val = (r+1)*fv;
@@ -325,14 +327,14 @@ void boxBlur_4 (DummyBitmap& scl, DummyBitmap& tcl, int w, int h, int r) {
 
 
 void gaussBlur_4 (DummyBitmap& scl, DummyBitmap& tcl, int w, int h, int r) {
-    int* bxs = boxesForGauss(r, 3);
+    int* bxs = boxesForGauss(static_cast<float>(r), 3);
     boxBlur_4 (scl, tcl, w, h, (bxs[0]-1)/2);
     boxBlur_4 (tcl, scl, w, h, (bxs[1]-1)/2);
     boxBlur_4 (scl, tcl, w, h, (bxs[2]-1)/2);
     delete[] bxs;
 }
 uint8_t *prevBuf = 0;
-int prevSize=0;
+unsigned int prevSize=0;
 
 void BlurCleanup() {
     delete[] prevBuf;
@@ -439,7 +441,7 @@ Gdiplus::Color StringToColor(const std::string& str) {
         std::vector<std::string> tokens;
         IuStringUtils::Split(str.substr(5, str.length()-5 ),",", tokens,4);
         if ( tokens.size() == 4 ) {
-            return Gdiplus::Color(round(atof(tokens[3].c_str())*255), atoi(tokens[0].c_str()), atoi(tokens[1].c_str()), atoi(tokens[2].c_str()));
+            return Gdiplus::Color(static_cast<BYTE>(round(atof(tokens[3].c_str())*255)), atoi(tokens[0].c_str()), atoi(tokens[1].c_str()), atoi(tokens[2].c_str()));
         }
     } else if ( str.substr(0,4) == "rgb" && str.length() >= 13 ) {
         std::vector<std::string> tokens;
